+++ package.json
@@ -23,7 +23,6 @@
     "@trpc/next": "^10.0.0",
     "@trpc/react-query": "^10.0.0",
     "@trpc/server": "^10.0.0",
-    "express": "^4.18.2",
     "googleapis": "^109.0.1",
     "next": "13.0.2",
     "next-auth": "^4.15.1",
@@ -35,7 +34,6 @@
     "zod": "^3.18.0"
   },
   "devDependencies": {
-    "@types/express": "^4.17.16",
     "@types/node": "^18.0.0",
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
